fix: add caps to delay startup of mcu
Sometimes the mcu does not boot correctly when usb gets connected while the system is off - most likely because the flash does not get enough power quick enough. Adding a startup delay for the 3V3 regulator and the mcu itself helps in my tests. This delays the boot by about half a second.